move processor to storage layer
