+++ snuba/datasets/storages/errors_common.py
@@ -21,6 +21,9 @@ from snuba.query.expressions import Column, Literal
 from snuba.query.processors.arrayjoin_keyvalue_optimizer import (
     ArrayJoinKeyValueOptimizer,
 )
+from snuba.query.processors.empty_tag_condition_processor import (
+    EmptyTagConditionProcessor,
+)
 from snuba.query.processors.mapping_optimizer import MappingOptimizer
 from snuba.query.processors.mapping_promoter import MappingColumnPromoter
 from snuba.query.processors.prewhere import PrewhereProcessor
@@ -163,6 +166,7 @@ query_processors = [
     EventsBooleanContextsProcessor(),
     TypeConditionOptimizer(),
     MappingOptimizer("tags", "_tags_hash_map", "events_tags_hash_map_enabled"),
+    EmptyTagConditionProcessor(),
     ArrayJoinKeyValueOptimizer("tags"),
     PrewhereProcessor(
         prewhere_candidates,
