+++ snuba/datasets/storages/transactions.py
@@ -22,6 +22,9 @@ from snuba.query.processors.arrayjoin_keyvalue_optimizer import (
     ArrayJoinKeyValueOptimizer,
 )
 from snuba.query.processors.conditions_enforcer import ProjectIdEnforcer
+from snuba.query.processors.empty_tag_condition_processor import (
+    EmptyTagConditionProcessor,
+)
 from snuba.query.processors.mapping_optimizer import MappingOptimizer
 from snuba.query.processors.mapping_promoter import MappingColumnPromoter
 from snuba.query.processors.prewhere import PrewhereProcessor
@@ -113,6 +116,7 @@ storage = WritableTableStorage(
         HexIntColumnProcessor({"span_id"}),
         EventsBooleanContextsProcessor(),
         MappingOptimizer("tags", "_tags_hash_map", "tags_hash_map_enabled"),
+        EmptyTagConditionProcessor(),
         ArrayJoinKeyValueOptimizer("tags"),
         ArrayJoinKeyValueOptimizer("measurements"),
         ArrayJoinKeyValueOptimizer("span_op_breakdowns"),
